% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDC.R
\name{RDC}
\alias{RDC}
\title{Robust Distance Correlation for Sure Screening.}
\usage{
RDC(x, y, c4 = 2, c2 = 2, auto = FALSE)
}
\arguments{
\item{x}{Predictors, a matrix of dimensions n * p.}

\item{y}{Response variables, a vector of length n.}

\item{c4}{Truncation parameter, which will further be used to solve for the
robustification parameter \eqn{\tau}. See the paper for details. Default is 2.}

\item{c2}{Truncation parameter, which will further be used to solve for the
robustification parameter \eqn{\tau \prime}. See the paper for details. Default is 2.}

\item{auto}{If TRUE, then the parameter c2 would automatically be tuned
based on c4. Default is FALSE.}
}
\value{
A vector of screening statistics for each predictor.
}
\description{
This function implements a Huber-type robust distance correlation measure
(called "Robust DC") for sure screening in ultrahigh-dimensional heavy-tailed
data.
}
\examples{
\dontrun{

set.seed(1234)
## parameter setting
n <- 50
p <- 500
s <- 2
x <- matrix(NA,n,p)
true.ind <- round(seq(1,p,length.out=s))
noise.ind <- (1:p)[-true.ind]
beta <- rep(0,p)
beta[true.ind] <- c(1,-1)
## simulate data
for(j in 1:p){
  x[,j] <- rlnorm(n)
}
y <- x\%*\%beta + rnorm(n)
## Robust DC
rdc.stat <- RDC(x,y)

}
}
