% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDC.R
\name{RDC}
\alias{RDC}
\title{Robust Distance Correlation for Sure Screening.}
\usage{
RDC(x, y, c1 = 0.5)
}
\arguments{
\item{x}{Predictors, a matrix of dimensions n * p.}

\item{y}{Response variables, a vector of length n.}

\item{c1}{Truncation parameter, which will further be used to sovle for the
robustification parameter \code{gamma}. See the paper for details. The larger
c1, the more truncation. Default is 0.5.}
}
\value{
A vector of screening statistics for each predictor.
}
\description{
This function implements a Huber-type robust distance correlation measure
(called "Robust DC") for sure screening in ultrahigh-dimensional heavy-tailed
data.
}
\examples{
\dontrun{

set.seed(1234)
## parameter setting
n <- 50
p <- 500
s <- 2
x <- matrix(NA,n,p)
true.ind <- round(seq(1,p,length.out=s))
noise.ind <- (1:p)[-true.ind]
beta <- rep(0,p)
beta[true.ind] <- c(1,-1)
## simulate data
for(j in 1:p){
  x[,j] <- rlnorm(n)
}
y <- x\%*\%beta + rnorm(n)
## Robust DC
rdc.stat <- RDC(x,y,c1=0.5)

}
}
